add subtraits `ModusConstant` and `ModusVariable`
This should make it easier to work with the <C, V> generics.
Also, add default concrete types for the `C`s and `V`s in the main
structs so users wouldn't need to deal with it at all. However, there
may be some benefit to keeping it, as I have done, because it reduces a
lot of potential duplication in the main algorithms, e.g. in the resolution.