only convert escape chars when translating to IR
An alternative to this approach would be to store the source span in the
other term types as well, but I've generally taken the approach to only
store the source span where it's important to have them.
e.g. it is unlikely that an error will occur in SLD where the constant
term used is the key part that caused that error.